Assignment 42 comments: trying to fix

+++ app/controllers/comments_controller.rb
@@ -4,29 +4,62 @@ class CommentsController < ApplicationController
   before_action :authorize_user, only: [:destroy]
 
   def create
-    @post = Post.find(params[:post_id])
-    comment = @post.comments.new(comment_params)
-    comment.user = current_user
 
-    if comment.save
-      flash[:notice] = "Comment saved successfully."
-      redirect_to [@post.topic, @post]
-    else
-      flash[:alert] = "Comment failed to save."
-      redirect_to [@post.topic, @post]
+    if params.include?(:post_id)
+      @post = Post.find(params[:post_id])
+      comment = @post.comments.new(comment_params)
+      comment.user = current_user
+
+      if comment.save
+        flash[:notice] = "Comment saved successfully."
+        redirect_to [@post.topic, @post]
+      elsif
+        flash[:alert] = "Comment failed to save."
+        redirect_to [@post.topic, @post]
+      end
+
+    else params.include?(:topic_id)
+      @topic = Topic.find(params[:topic_id])
+      comment = @topic.comments.new(comment_params)
+      comment.user = current_user
+
+      if comment.save
+        flash[:notice] = "Comment saved successfully."
+        redirect_to [@topic]
+      else
+        flash[:alert] = "Comment failed to save."
+        redirect_to [@topic]
+      end
     end
+
   end
 
   def destroy
-    @post = Post.find(params[:post_id])
-    comment = @post.comments.find(params[:id])
 
-    if comment.destroy
-      flash[:notice] = "Comment was deleted successfully."
-      redirect_to [@post.topic, @post]
-    else
-      flash[:alert] = "Comment couldn't be deleted. Try again."
-      redirect_to [@post.topic, @post]
+    if params.include?(:post_id)
+      @post = Post.find(params[:post_id])
+      comment = @post.comments.find(params[:id])
+
+      if comment.destroy
+        flash[:notice] = "Comment was deleted successfully."
+        redirect_to [@post.topic, @post]
+      else
+        flash[:alert] = "Comment couldn't be deleted. Try again."
+        redirect_to [@post.topic, @post]
+      end
+    end
+
+    if params.include?(:topic_id)
+      @topic = Topic.find(params[:topic_id])
+      comment = @topic.comments.find(params[:id])
+
+      if comment.destroy
+        flash[:notice] = "Comment was deleted successfully."
+        redirect_to [@topic]
+      else
+        flash[:alert] = "Comment couldn't be deleted. Try again."
+        redirect_to [@topic]
+      end
     end
   end
 
@@ -40,7 +73,12 @@ class CommentsController < ApplicationController
     comment = Comment.find(params[:id])
     unless current_user == comment.user || current_user.admin?
       flash[:alert] = "You do not have permission to delte a comment."
-      redirect_to [comment.post.topic, comment.post]
+
+      if params.include?(:post_id)
+        redirect_to [comment.post.topic, comment.post]
+      elsif params.include?(:topic_id)
+        redirect_to [comment.topic, comment.topic]
+      end
     end
   end
 end
